+++ resources/views/documents/print.blade.php
@@ -91,7 +91,7 @@
                     <div class="row"><span class="k">Reference No.</span><span class="v">{{ $document->reference_no }}</span></div>
                 @endif
                 @if($document->fund)
-                    <div class="row"><span class="k">Fund</span><span class="v">{{ $document->fund->name }} ({{ $document->fund->code }})</span></div>
+                    <div class="row"><span class="k">Fund</span><span class="v">{{ $document->fund->name }}</span></div>
                 @endif
                 @if($document->amount !== null)
                     <div class="row"><span class="k">Amount</span><span class="v">₱{{ number_format($document->amount, 2) }}</span></div>